pyproject.toml
[tool.poetry]
name = "app"
version = "0.2.1"
description = "Demo App"
authors = []

[tool.poetry.dependencies]
python = ">=3.9.0,<4"
fastapi =  "~0.66.0"
opentelemetry-instrumentation-fastapi = "~0.22-beta.0"
opentelemetry-sdk = "~1.3.0"
uvicorn = { version = "~0.13.4", extras = ["standard"] }
opentelemetry-exporter-otlp = "~1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
